+++ src/clj/lambdacd/event_bus.clj
@@ -7,8 +7,16 @@
     (assoc ctx :event-publisher   publisher-ch
                :event-publication publication)))
 
+
+(defmacro publish! [ctx topic payload]
+  `(async/>! (:event-publisher ~ctx) {:topic ~topic :payload ~payload}))
+
+(defmacro publish!! [ctx topic payload]
+  `(async/>!! (:event-publisher ~ctx) {:topic ~topic :payload ~payload}))
+
 (defn publish [ctx topic payload]
-  (async/>!! (:event-publisher ctx) {:topic topic :payload payload}))
+  "DEPRECATED, will be removed in subsequent versions. use publish!! or publish! instead"
+  (publish!! ctx topic payload))
 
 (defn subscribe [ctx topic]
   (let [result-ch (async/chan)]